Fix path encoding issues (#1153) (#2654)
* Test Unicode functionality

* Lua support for UTF-8 on Windows

* Premake support for UTF-8

+++ src/host/path_isabsolute.c
@@ -23,7 +23,7 @@ int do_absolutetype(const char* path)
 		path++;
 	if (path[0] == '/' || path[0] == '\\')
 		return JOIN_ABSOLUTE;
-	if (isalpha(path[0]) && path[1] == ':')
+	if (isalpha((int)path[0]) && path[1] == ':')
 		return JOIN_ABSOLUTE;
 
 	// $(foo) and %(foo)
